consertado erros no docs e retirando classes que foram incorporadas nas entidades
