+++ housinapi/src/main/java/com/dcx/pas1/housinapi/controller/AnuncioController.java
@@ -2,16 +2,15 @@ package com.dcx.pas1.housinapi.controller;
 
 import com.dcx.pas1.housinapi.model.Anuncio;
 import com.dcx.pas1.housinapi.repository.AnunciosRepository;
+import io.swagger.annotations.ApiOperation;
+import io.swagger.annotations.ApiResponse;
+import io.swagger.annotations.ApiResponses;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
-
-import io.swagger.annotations.ApiOperation;
-import io.swagger.annotations.ApiResponse;
-import io.swagger.annotations.ApiResponses;
 
 import javax.persistence.EntityNotFoundException;
 import javax.servlet.http.HttpServletResponse;
@@ -30,7 +29,7 @@ public class AnuncioController {
     @ApiResponse(code = 200, message = "Retorna os anúncios cadastrados"),
     @ApiResponse(code = 204, message = "Não há anúncios cadastrados")
 })
-  @ApiOperation(value = "Retorna todas os anúncios cadastrados no sistema.", response = Iterable.class, produces = json)
+  @ApiOperation(value = "Retorna todas os anúncios cadastrados no sistema.", response = Anuncio.class, responseContainer = "List", produces = json)
   @GetMapping
   @ResponseStatus(HttpStatus.OK)
   public ResponseEntity<?> getAnuncios(){
@@ -38,7 +37,7 @@ public class AnuncioController {
     return !anuncios.isEmpty() ? ResponseEntity.ok(anuncios) : ResponseEntity.noContent().build();
   }
 
-	@ApiOperation(value = "Cria um novo anúncio", response = Iterable.class, produces = json)
+    @ApiOperation(value = "Cria um novo anúncio", response = Anuncio.class, produces = json, consumes = json)
   @PostMapping
   @ResponseStatus(HttpStatus.CREATED)
 	public ResponseEntity<Anuncio> criar(@Valid @RequestBody Anuncio anuncio, HttpServletResponse response) {
@@ -55,7 +54,7 @@ public class AnuncioController {
     @ApiResponse(code = 200, message = "Busca anúncio específico pelo código"),
     @ApiResponse(code = 204, message = "Não há anúncios cadastrados")
 })
-	@ApiOperation(value = "Busca anúncio específico pelo código", response = Iterable.class, produces = json)
+  @ApiOperation(value = "Busca anúncio específico pelo código", response = Anuncio.class, produces = json)
 	@GetMapping("/{codigo}")
   @ResponseStatus(HttpStatus.FOUND)
   public Anuncio buscarPeloCodigo(@PathVariable Long codigo) {
@@ -66,14 +65,14 @@ public class AnuncioController {
   @ApiResponses(value = {
     @ApiResponse(code = 200, message = "Deleta anúncio específico pelo código")
   })
-  @ApiOperation(value = "Deleta anúncio específico pelo código", response = Iterable.class, produces = json)
+  @ApiOperation(value = "Deleta anúncio específico pelo código")
   @DeleteMapping("/{codigo}")
   @ResponseStatus(HttpStatus.NO_CONTENT)
   public void deletaPeloCodigo(@PathVariable Long codigo){
     this.anuncioRepository.deleteById(codigo);
   }
 
-  @ApiOperation(value = "Atualiza os dados de um anúncio", response = Iterable.class, produces = json)
+    @ApiOperation(value = "Atualiza os dados de um anúncio", response = Anuncio.class, produces = json)
   @PutMapping("/{codigo}")
   public ResponseEntity<Anuncio> atualizaAnuncio(@PathVariable Long codigo, @Valid @RequestBody Anuncio anuncio){
     Anuncio anunciosalvo = this.anuncioRepository.findById(codigo)
