+++ housinapi/src/main/java/com/dcx/pas1/housinapi/model/Casa.java
@@ -1,7 +1,5 @@
 package com.dcx.pas1.housinapi.model;
 
-import io.swagger.annotations.ApiModelProperty;
-
 import javax.persistence.*;
 import javax.validation.constraints.NotNull;
 
@@ -9,41 +7,51 @@ import javax.validation.constraints.NotNull;
 @Table(name = "casa")
 public class Casa {
 
-    @ApiModelProperty(value = "Código identificador da casa(anúncio)", example = "0")
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private long codigoCasa;
 
-    @ApiModelProperty(value = "Quantos moradores há na casa", example = "2")
     @NotNull
     private int quantMoradores;
 
-    @ApiModelProperty(value = "Especificações da casa, como quantidade e cômodos disponíveis")
     @NotNull
-    @OneToOne
-    @JoinColumn(name = "codigo_especificacoes")
-    private Especificacoes especificacoes;
+    private int quantQuartos;
 
-    @ApiModelProperty(value = "Objeto de endereço da casa", example = "Rua da Alegria, 349, Mamanguape, Centro")
     @NotNull
-    @Embedded
-    private Endereco endereco;
+    private int quantBanheiros;
 
-    public Especificacoes getEspecificacoes() {
-        return especificacoes;
-    }
+    @NotNull
+    private boolean hasGaragem;
 
-    public void setEspecificacoes(Especificacoes especificacoes) {
-        this.especificacoes = especificacoes;
-    }
+    @NotNull
+    private boolean hasFogao;
 
-    public Endereco getEndereco() {
-        return endereco;
-    }
+    @NotNull
+    private boolean hasGeladeira;
 
-    public void setEndereco(Endereco endereco) {
-        this.endereco = endereco;
-    }
+    @NotNull
+    private boolean hasTv;
+
+    @NotNull
+    private boolean hasInternet;
+
+    @NotNull
+    private boolean hasMicroondas;
+
+    @NotNull
+    private String logradouro;
+
+    @NotNull
+    private String numero;
+
+    @NotNull
+    private String bairro;
+
+    @NotNull
+    private String cep;
+
+    @NotNull
+    private String cidade;
 
     public int getQuantMoradores() {
         return quantMoradores;
@@ -61,5 +69,107 @@ public class Casa {
         this.codigoCasa = codigoCasa;
     }
 
-    
+    public int getQuantQuartos() {
+        return quantQuartos;
+    }
+
+    public void setQuantQuartos(int quantQuartos) {
+        this.quantQuartos = quantQuartos;
+    }
+
+    public int getQuantBanheiros() {
+        return quantBanheiros;
+    }
+
+    public void setQuantBanheiros(int quantBanheiros) {
+        this.quantBanheiros = quantBanheiros;
+    }
+
+    public boolean isHasGaragem() {
+        return hasGaragem;
+    }
+
+    public void setHasGaragem(boolean hasGaragem) {
+        this.hasGaragem = hasGaragem;
+    }
+
+    public boolean isHasFogao() {
+        return hasFogao;
+    }
+
+    public void setHasFogao(boolean hasFogao) {
+        this.hasFogao = hasFogao;
+    }
+
+    public boolean isHasGeladeira() {
+        return hasGeladeira;
+    }
+
+    public void setHasGeladeira(boolean hasGeladeira) {
+        this.hasGeladeira = hasGeladeira;
+    }
+
+    public boolean isHasTv() {
+        return hasTv;
+    }
+
+    public void setHasTv(boolean hasTv) {
+        this.hasTv = hasTv;
+    }
+
+    public boolean isHasInternet() {
+        return hasInternet;
+    }
+
+    public void setHasInternet(boolean hasInternet) {
+        this.hasInternet = hasInternet;
+    }
+
+    public boolean isHasMicroondas() {
+        return hasMicroondas;
+    }
+
+    public void setHasMicroondas(boolean hasMicroondas) {
+        this.hasMicroondas = hasMicroondas;
+    }
+
+    public String getLogradouro() {
+        return logradouro;
+    }
+
+    public void setLogradouro(String logradouro) {
+        this.logradouro = logradouro;
+    }
+
+    public String getNumero() {
+        return numero;
+    }
+
+    public void setNumero(String numero) {
+        this.numero = numero;
+    }
+
+    public String getBairro() {
+        return bairro;
+    }
+
+    public void setBairro(String bairro) {
+        this.bairro = bairro;
+    }
+
+    public String getCep() {
+        return cep;
+    }
+
+    public void setCep(String cep) {
+        this.cep = cep;
+    }
+
+    public String getCidade() {
+        return cidade;
+    }
+
+    public void setCidade(String cidade) {
+        this.cidade = cidade;
+    }
 }
